Fixed bug in updated contributions in ELO store

+++ include/orchestrator.hpp
@@ -4,7 +4,7 @@
 #include <unistd.h>
 
 // define engine types
-typedef int32_t cid;
+typedef size_t cid;
 typedef float elo;
 
 // define manager paramaters
@@ -19,8 +19,8 @@ const float ELO_N_SCALE = 400.0;
 const float ELO_EXP_BASE = 10.0;
 
 // define strucutre parameters
-const int32_t ELO_STORE_SIZE = 5000;
-const int32_t CONTRIBUTION_STORE_SIZE = 100000;
-const int32_t INITIAL_CONTRIBUTION_COUNT = 0;
+const size_t ELO_STORE_SIZE = 5000;
+const size_t CONTRIBUTION_STORE_SIZE = 100000;
+const size_t INITIAL_CONTRIBUTION_COUNT = 0;
 
 #endif
